alterações na controller user
mudanças no tratamento de erros e img default na !req.file

+++ backend/controllers/UserController.js
@@ -28,10 +28,19 @@ module.exports = class UserController {
             return isNaN(number) ? value : number
         }
 
+        let image = req.file
+
+        if (!image) {
+            image = 'default.jpg'
+        } else {
+            image = image.filename
+        }
+
         const userParse = {
             name: req.body.name,
             email: req.body.email,
             phone: toNumber(req.body.phone),
+            img: image,
             password: req.body.password,
             confirmpassword: req.body.confirmpassword,
         }
@@ -40,6 +49,7 @@ module.exports = class UserController {
             name: z.string().min(3, { message: 'O nome precisa ter pelo menos 3 caracteres, por favor verifique o que foi digitado' }).max(25, { message: 'O nome pode ter no máximo 25 caracteres ,por favor verifique o que foi digitado' }),
             email: z.string().email({ message: 'Endereço de email inválido' }),
             phone: z.number({ message: 'Número com formato digitado inválido' }).min(11, { message: 'Número de telefone digitado inválido, certifique de colocar o ddd da sua localidade' }),
+            img: z.string(),
             password: z.string().min(6, { message: 'Senha muito curta ,mínimo 6 characteres , por favor faça outra' }).max(28, { message: 'Senha muito grande , por favor faça outra' }),
         })
 
@@ -59,6 +69,7 @@ module.exports = class UserController {
 
         if (userParse.password !== userParse.confirmpassword) {
             res.status(422).json({ message: 'A senha e a sua confirmação não são iguais , por favor verifique o que foi digitado' })
+            return
         }
 
         const UserExists = await User.findOne({ email: userParse.email })
@@ -78,6 +89,7 @@ module.exports = class UserController {
             name: userParse.name,
             email: userParse.email,
             phone: userParse.phone,
+            img: userParse.img,
             password: PasswordHash,
             primaryLogin,
             token
@@ -224,7 +236,7 @@ module.exports = class UserController {
                 const newToken = GenerateRandomFourDigitNumber()
                 await User.findOneAndUpdate(
                     { _id: userDb._id },
-                    { $set: { password: PasswordHash ,token : newToken} }
+                    { $set: { password: PasswordHash, token: newToken } }
                 )
                 await CreateUserToken(userDb, req, res)
             } catch (erro) {
@@ -296,10 +308,20 @@ module.exports = class UserController {
             return value
         }
 
+        let image = req.file
+        console.log(`image : ${JSON.stringify(image)}`)
+        if (!image) {
+            image = 'default.jpg'
+            console.log('entrou aqui')
+        } else {
+            image = image.filename
+        }
+
         const userParse = {
             name: req.body.name,
             email: req.body.email,
             phone: toNumber(req.body.phone),
+            img: image,
             password: req.body.password,
             confirmpassword: req.body.confirmpassword,
             receiveremail: toBoolean(req.body.receiveremail)
@@ -308,7 +330,8 @@ module.exports = class UserController {
         const userSchema = z.object({
             name: z.string().min(3, { message: 'O nome precisa ter pelo menos 3 caracteres, por favor verifique o que foi digitado' }).max(25, { message: 'O nome pode ter no máximo 25 caracteres ,por favor verifique o que foi digitado' }),
             email: z.string().email({ message: 'Endereço de email inválido' }),
-            phone: z.number({ message: 'Número com formato digitado inválido' }).min(1000000000, { message: 'Número de telefone digitado inválido, certifique de colocar o ddd da sua localidade' }).max(99999999999, { message: 'Número de telefone digitado inválido, certifique dos dados digitados' }),
+            phone: z.number({ message: 'Número com formato digitado inválido' }).min(11, { message: 'Número de telefone digitado inválido, certifique de colocar o ddd da sua localidade' }),
+            img: z.string(),
             password: z.string().min(6, { message: 'Senha muito curta ,mínimo 6 characteres , por favor faça outra' }).max(28, { message: 'Senha muito grande , por favor faça outra' }),
             receiveremail: z.boolean({ message: 'Tipo de dado incorreto para o recebimento de e-mail , somente permitido true ou false' })
         })
@@ -344,7 +367,7 @@ module.exports = class UserController {
         }
 
         if (req.file) {
-            if (userDb.img) {
+            if (userDb.img !== 'default.jpg' && userDb.img) {
                 const imagePath = path.join(__dirname, '..', 'public', 'images', 'users', userDb.img);
                 fs.unlink(imagePath, (err) => {
                     if (err) {
@@ -353,6 +376,16 @@ module.exports = class UserController {
                 })
             }
             userDb.img = req.file.filename
+        } else {
+            if (userDb.img !== 'default.jpg' && userDb.img) {
+                const imagePath = path.join(__dirname, '..', 'public', 'images', 'users', userDb.img);
+                fs.unlink(imagePath, (err) => {
+                    if (err) {
+                        console.error('Erro ao excluir a imagem:', err);
+                    }
+                })
+            }
+            userDb.img = userParse.img
         }
 
         userDb.name = userParse.name
